String library fixes

+++ include/StringEx.h
@@ -43,17 +43,17 @@ extern "C" {
 extern LIBRARY_EXPORT wchar_t* strtowstr(const char* str);
 extern LIBRARY_EXPORT char* strfromwstr(const wchar_t* wstr);
 
-extern LIBRARY_EXPORT char* strfromint(size_t num);
+extern LIBRARY_EXPORT char* strfromint(long num);
 extern LIBRARY_EXPORT char* strfromdouble(double num);
 
 extern LIBRARY_EXPORT char* strreverse(char* ptr);
-extern LIBRARY_EXPORT char* strsegmentreverse(char* str, size_t start, size_t term);
+extern LIBRARY_EXPORT char* strsegmentreverse(char* str, long start, long term);
 
 extern LIBRARY_EXPORT long strindexofsubstr(const char *str, const char* substr);
 extern LIBRARY_EXPORT long strindexofchar(char* str, const char ch);
 
-extern LIBRARY_EXPORT size_t strcountsubstr(const char *str, const char* substr);
-extern LIBRARY_EXPORT size_t strcountchar(char* str, const char ch);
+extern LIBRARY_EXPORT long strcountsubstr(const char *str, const char* substr);
+extern LIBRARY_EXPORT long strcountchar(char* str, const char ch);
 
 extern LIBRARY_EXPORT char* strtolower(char* str);
 extern LIBRARY_EXPORT char* strtoupper(char* str);
@@ -64,25 +64,25 @@ extern LIBRARY_EXPORT char* stralltrim(char* str);
 
 extern LIBRARY_EXPORT char* strremsubstrfirst(char* str, const char* substr);
 extern LIBRARY_EXPORT char* strremsubstrall(char* str, const char* substr);
-extern LIBRARY_EXPORT char* strremsubstrat(char* str, size_t pos, size_t len);
+extern LIBRARY_EXPORT char* strremsubstrat(char* str, long pos, long len);
 
 extern LIBRARY_EXPORT char* strremcharfirst(char* str, const char oldchar);
 extern LIBRARY_EXPORT char* strremcharall(char* str, const char oldchar);
-extern LIBRARY_EXPORT char* strremcharat(char* str, size_t pos);
+extern LIBRARY_EXPORT char* strremcharat(char* str, long pos);
 
 extern LIBRARY_EXPORT char* strrepsubstrfirst(char* str, const char* oldsubstr, const char* newsubstr);
 extern LIBRARY_EXPORT char* strrepsubstrall(char* str, const char* oldsubstr, const char* newsubstr);
 
 extern LIBRARY_EXPORT char* strrepcharfirst(char* str, const char oldchar, const char newchar);
 extern LIBRARY_EXPORT char* strrepcharall(char* str, const char oldchar, const char newchar);
-extern LIBRARY_EXPORT char* strrepcharat(char* str, const char newchar, size_t pos);
+extern LIBRARY_EXPORT char* strrepcharat(char* str, const char newchar, long pos);
 
 extern LIBRARY_EXPORT void strsplitkeyvalue(const char* str, const char* delimiter, char **key, char **value);
-extern LIBRARY_EXPORT char** strsplitsubstr(const char* str, const char* delimiter, size_t *numsubstr);
-extern LIBRARY_EXPORT char** strsplitchar(const char* str, const char delimiter, size_t *numsubstr);
+extern LIBRARY_EXPORT char** strsplitsubstr(const char* str, const char* delimiter, long *numsubstr);
+extern LIBRARY_EXPORT char** strsplitchar(const char* str, const char delimiter, long *numsubstr);
 extern LIBRARY_EXPORT char* strjoinwithsubstr(const char **strlist, const char* delimiter);
 extern LIBRARY_EXPORT char* strjoinwithchar(const char** strlist, const char delimiter);
-extern LIBRARY_EXPORT void  strfreelist(char** strlist, size_t numsubstr);
+extern LIBRARY_EXPORT void  strfreelist(char** strlist, long numsubstr);
 
 #ifdef __cplusplus
 }
